Convert hashes to base + real.

+++ fuzz/mk_sha1_fuzz.h
@@ -2,10 +2,7 @@
 #define mk_include_guard_sha1_fuzz
 
 
-#include <stddef.h> /* size_t */
-
-
-void mk_sha1_fuzz(unsigned char const* data, size_t size);
+void mk_sha1_fuzz(unsigned char const* data, int size);
 
 
 #endif
